Revert both changes. (#840)
* Revert "Improve queries 2nd attampt. (#839)"

This reverts commit c60e7ba58985a6ce599df29bf7463aa9263dd69a.

* Revert "Make atlas happy (#838)"

This reverts commit 56bfbe6c9626e5fec69c66162d7365dcaec58a40.

+++ spark/so/ent/signingkeyshare_extension.go
@@ -57,7 +57,7 @@ func (keyshare *SigningKeyshare) MarshalProto() *pb.SigningKeyshare {
 
 	return &pb.SigningKeyshare{
 		OwnerIdentifiers: ownerIdentifiers,
-		Threshold:        uint32(keyshare.MinSigners),
+		Threshold:        keyshare.MinSigners,
 	}
 }
 
@@ -151,7 +151,7 @@ func GetKeyPackage(ctx context.Context, config *so.Config, keyshareID uuid.UUID)
 		SecretShare:  keyshare.SecretShare,
 		PublicShares: keyshare.PublicShares,
 		PublicKey:    keyshare.PublicKey,
-		MinSigners:   uint32(keyshare.MinSigners),
+		MinSigners:   keyshare.MinSigners,
 	}
 
 	return keyPackage, nil
@@ -173,7 +173,7 @@ func GetKeyPackages(ctx context.Context, config *so.Config, keyshareIDs []uuid.U
 			SecretShare:  keyshare.SecretShare,
 			PublicShares: keyshare.PublicShares,
 			PublicKey:    keyshare.PublicKey,
-			MinSigners:   uint32(keyshare.MinSigners),
+			MinSigners:   keyshare.MinSigners,
 		}
 	}
